Fixing bugs with DriverAssist and protocol to Rio.
DriverAssist.ino: Placing quotes around the trustMe letters, so they are stored as characters, not variables. Removing old autoPilotAvailable variable. Removing early conversion of values to 0-100.

rioCommunications.h: Fixing bounds to allow negative values. Adding number of characters parameter to addLeadingZeros so it will begin writing to the buffer after the data that is already written. Removing unused number_of_digits function. Correctin buffer size. Fixing the bytes_written to be the current location of the buffer, instead of adding that to the existing location. Fixing prints to writes so the values are written as characters instead of printing the underlying number value of the characters.